VAB-1000 u-boot source v2.01.00 release changes on top of upstream
Changes are provided as is based on the released u-boot source
bundle.

+++ include/jffs2/load_kernel.h
@@ -46,8 +46,10 @@ struct part_info {
 	struct list_head link;
 	char *name;			/* partition name */
 	u8 auto_name;			/* set to 1 for generated name */
-	u32 size;			/* total size of the partition */
-	u32 offset;			/* offset within device */
+	//u32 size;			/* total size of the partition */
+	u64 size;
+	//u32 offset;			/* offset within device */
+	u64 offset;
 	void *jffs2_priv;		/* used internaly by jffs2 */
 	u32 mask_flags;			/* kernel MTD mask flags */
 	u32 sector_size;		/* size of sector */
@@ -58,7 +60,8 @@ struct mtdids {
 	struct list_head link;
 	u8 type;			/* device type */
 	u8 num;				/* device number */
-	u32 size;			/* device size */
+	//u32 size;			/* device size */
+	u64 size;
 	char *mtd_id;			/* linux kernel device id */
 };
